+++ src/MicroElements.Metadata/Metadata/IPropertyContainer.cs
@@ -67,13 +67,4 @@ namespace MicroElements.Metadata
         /// </summary>
         void Clear();
     }
-
-    public interface IPropertyContainer<TSchema> : IPropertyContainer, IKnownPropertySet<TSchema>, IPropertySet
-        where TSchema : IPropertySet, new()
-    {
-        TSchema Schema { get; }
-
-        /// <inheritdoc />
-        IEnumerable<IProperty> IPropertySet.GetProperties() => Schema.GetProperties();
-    }
 }